add shape & cat

+++ lib/shape.cc
@@ -9,7 +9,7 @@ torch::Tensor shape_t(const torch::Tensor &x_t)
         dim_tmp[i] = (int64_t)(a[i]);
     }
     c10::IntArrayRef a_dim (size);
-    auto options = torch::TensorOptions().dtype(torch::kInt64);
+    auto options = torch::TensorOptions().dtype(torch::kInt32);
     torch::Tensor z_t = torch::from_blob(dim_tmp, a_dim, options);
     return z_t.clone();
 }
